refactor so we're using the exact same throttles for both apis
